tests fixed

+++ tests/unit/auxiliaries/math/tools/main.cpp
@@ -190,7 +190,7 @@ int main()
   errors += testMatrixAccess1D<uint8_t, uint8_t, CV_8U>();
   errors += testMatrixAccess1D<uint16_t, uint16_t, CV_16U>();
   errors += testMatrixAccess1D<float, uint32_t, CV_32F>();
-  errors += testMatrixAccess1D<double, uint64_t, CV_32F>();
+  errors += testMatrixAccess1D<double, uint64_t, CV_64F>();
 
   std::cout << "test no " << test_number++ << std::endl;
   if (cedar::aux::math::normalizeAngle(5.9) <= -cedar::aux::math::pi
